+++ Sources/OnboardingKit/PageView/PageView.swift
@@ -13,10 +13,13 @@ import Combine
 /// overlays and gestures to mimic a `.page` `TabView`.
 ///
 /// You can set it up with a fixed set of `pages`, or with a
-/// set of `items` to which you apply a dynamic page builder.
+/// set of `items` and a page view builder.
+///
+/// The view supports arrow navigation and tapping the edges
+/// to navigate between pages.
 struct PageView<PageViewType: View>: View {
     
-    /// Create a page view with a set of pre-built pages.
+    /// Create a page view with a list of page views.
     init(
         pages: [PageViewType],
         currentPageIndex: Binding<Int>,
@@ -28,8 +31,9 @@ struct PageView<PageViewType: View>: View {
         self.pageIndicatorDisplayMode = pageIndicatorDisplayMode
         self.pageIndicatorStyle = pageIndicatorStyle
     }
-    
-    /// Create a page view with items and a page builder.
+
+    /// Create a page view with a list of values, and a page
+    /// view builder that is called for each value.
     init<Model>(
         pages: [Model],
         currentPageIndex: Binding<Int>,
@@ -42,12 +46,14 @@ struct PageView<PageViewType: View>: View {
         self.pageIndicatorDisplayMode = pageIndicatorDisplayMode
         self.pageIndicatorStyle = pageIndicatorStyle
     }
-    
+
     private var currentPageIndex: Binding<Int>
     private let pageIndicatorDisplayMode: PageIndicator.DisplayMode
     private let pageIndicatorStyle: PageIndicatorStyle
-    private let pages: [PageViewType]
-    
+    private var pages: [PageViewType]
+
+    @Environment(\.layoutDirection) var layoutDirection
+
     var body: some View {
         GeometryReader { geo in
             ZStack(alignment: .bottom) {
@@ -55,21 +61,74 @@ struct PageView<PageViewType: View>: View {
                     ScrollView(.horizontal) {
                         pageStack(for: geo)
                     }
-                    .onReceive(Just(currentPageIndex), perform: {
-                        setPageIndex(to: $0.wrappedValue, with: scroll)
-                    })
-                    .overlay(swipeLayer(for: scroll))
+                    .prefersScrollAdjustments(
+                        for: scroll,
+                        layoutDirection: layoutDirection,
+                        showPrevious: { showPreviousPage() },
+                        showNext: { showNextPage() }
+                    )
+                    .onReceive(Just(currentPageIndex)) { index in
+                        withAnimation {
+                            scroll.scrollTo(index.wrappedValue)
+                        }
+                    }
                     .overlay(bugfixLayer(for: scroll))
                 }
                 pageIndicator
-                
             }
         }
     }
 }
 
+private extension View {
+
+    @ViewBuilder
+    func prefersScrollAdjustments(
+        for scroll: ScrollViewProxy,
+        layoutDirection: LayoutDirection,
+        showPrevious: @escaping () -> Void,
+        showNext: @escaping () -> Void
+    ) -> some View {
+        #if os(iOS) || os(watchOS)
+        if #available(iOS 16.0, watchOS 9.0, *) {
+            self.scrollDisabled(true)
+        } else {
+            self
+        }
+        #else
+        if #available(macOS 13.0, tvOS 16.0, *) {
+            self.scrollDisabled(true)
+                .onMoveCommand { direction in
+                    switch direction.adjusted(for: layoutDirection) {
+                    case .left: showPrevious()
+                    case .right: showNext()
+                    default: break
+                    }
+                }
+        } else {
+            self
+        }
+        #endif
+    }
+}
+
+#if os(macOS) || os(tvOS)
+private extension MoveCommandDirection {
+
+    func adjusted(
+        for direction: LayoutDirection
+    ) -> Self {
+        switch self {
+        case .left: direction == .leftToRight ? .left : .right
+        case .right: direction == .leftToRight ? .right : .left
+        default: self
+        }
+    }
+}
+#endif
+
 private extension PageView {
-    
+
     @ViewBuilder
     var pageIndicator: some View {
         if shouldShowPageIndicator {
@@ -82,36 +141,49 @@ private extension PageView {
             EmptyView()
         }
     }
-    
+
+    // This is needed for the scroll view transition to play.
     func bugfixLayer(for scroll: ScrollViewProxy) -> some View {
         Text("\(currentPageIndex.wrappedValue)")
             .opacity(0)
     }
-    
-    func pageStack(for geo: GeometryProxy) -> some View {
+
+    func pageStack(
+        for geo: GeometryProxy
+    ) -> some View {
         HStack {
             ForEach(Array(pages.enumerated()), id: \.offset) {
                 $0.element
                     .id($0.offset)
                     .tag($0.offset)
                     .frame(width: geo.size.width, height: geo.size.height)
+                    .background(gestureLayer)
             }
         }
     }
     
-    func swipeLayer(for scroll: ScrollViewProxy) -> some View {
+    var gestureLayer: some View {
         #if os(tvOS)
-        swipeLayerView(for: scroll)
+        EmptyView()
         #else
-        swipeLayerView(for: scroll)
-            .onSwipeGesture(
-                left: { showNextPage(with: scroll) },
-                right: { showPreviousPage(with: scroll) })
+        HStack(spacing: 0) {
+            gestureLayerView(onTap: showPreviousPage)
+            gestureLayerView(onTap: showNextPage)
+        }
         #endif
     }
-    
-    func swipeLayerView(for scroll: ScrollViewProxy) -> some View {
+
+    func gestureLayerView(
+        onTap: @escaping () -> Void
+    ) -> some View {
         Color.white.opacity(0.00001)
+            .onSwipeGesture(
+                left: { showNextPage() },
+                right: { showPreviousPage() }
+            )
+            .simultaneousGesture(
+                TapGesture().onEnded(onTap)
+            )
     }
 }
 
@@ -124,21 +196,39 @@ private extension PageView {
         case .never: return false
         }
     }
+
+    func setPageIndex(to index: Int) {
+        currentPageIndex.wrappedValue = index
+    }
     
-    func setPageIndex(to index: Int, with scroll: ScrollViewProxy) {
-        withAnimation {
-            currentPageIndex.wrappedValue = index
-            scroll.scrollTo(index)
+    func showNextPage() {
+        guard currentPageIndex.wrappedValue < pages.count - 1 else { return }
+        setPageIndex(to: currentPageIndex.wrappedValue + 1)
+    }
+    
+    func showPreviousPage() {
+        guard currentPageIndex.wrappedValue > 0 else { return }
+        setPageIndex(to: currentPageIndex.wrappedValue - 1)
+    }
+}
+
+#Preview {
+
+    struct Preview: View {
+
+        @State var pageIndex = 0
+
+        var body: some View {
+            PageView(pages: Array(1...10), currentPageIndex: $pageIndex) { page in
+                VStack {
+                    Text("\(page)")
+                    HStack {
+                        Button("Next") { pageIndex += 1 }
+                    }
+                }
+            }
         }
     }
-    
-    func showNextPage(with scroll: ScrollViewProxy) {
-        guard currentPageIndex.wrappedValue < pages.count - 1 else { return }
-        setPageIndex(to: currentPageIndex.wrappedValue + 1, with: scroll)
-    }
-    
-    func showPreviousPage(with scroll: ScrollViewProxy) {
-        guard currentPageIndex.wrappedValue > 0 else { return }
-        setPageIndex(to: currentPageIndex.wrappedValue - 1, with: scroll)
-    }
+
+    return Preview()
 }